Add a test with messages for PostCSS and update CI configuration
Previously the PostCSS test wasn't actually running the linter as PostCSS is not a built in language in Atom so it was recognized as a Plain Text file. Add a spec that checks for a message to ensure linting is working properly.

Also updates the CI configuration to properly test this.