Stop merging stylelint-config-standard when enabled
Previously if the "Use standard" option was enabled the rules from that
were _merged_ into the rules coming from the project. Change the
behavior of this setting to act in the intended manner of allowing that
to be used when no other configuration was found. As such if this option
is enabled the "Disable when no config is found" option is forced to be
disabled when this is active and visa-versa.