Set proper configBaseDir for default configurations
When a user has failed to provide a configuration for their project and
we are generating one from the default `stylelint-config-standard`, the
`configBaseDir` needs to be set to the location where those rules can be
found. In this case that is the location of this package, not the
location of the project being linted as the user has no configuration of
their own and thus shouldn't be expected to have installed it in their
project.

Fixes https://github.com/AtomLinter/linter-stylelint/issues/357.