fix: use local stylelint for config resolution if available
Although the running of stylelint was changed to use the local project's
instance if available in #381, the instance used to determine the
configuration was unfortunately left as the bundled version. This made
no difference then, and made no difference when the bundled version was
still compatible with the local project one... but caused issues when
there is a mismatch.

This fixes it so we always use the stylelint instance local to the
project if there is one available, falling back to the bundled one if
one isn't available.

Fixes #440.